// src/app/modules/campaigns/core/models/campaign.enums.ts
export enum EUsagePeriodTypes {
  ALL_WEEK = "all_week",
  WEEKEND = "weekend",
  WEEKDAYS = "weekdays",
}

export enum EUsagePeriodTypesL {
  all_week = "Tüm hafta",
  weekend = "Hafta Sonu",
  weekdays = "Hafta İçi",
}

export enum EDiscountTypes {
  PERCENTAGE = "percentage",
  FIXED = "fix",
}

export enum EDiscountTypesL {
  percentage = "Yüzde",
  fix = "Sabit",
}

export enum ECampaignStatus {
  DRAFT = "draft",
  ACTIVE = "active",
  EXPIRED = "expired",
}

export enum ECampaignStatusL {
  draft = "Taslak",
  active = "Aktif",
  expired = "Günü Geçmiş",
}